+++ routingpy/routers/__init__.py
@@ -18,6 +18,7 @@ from ..client_base import options  # noqa: F401
 from ..exceptions import RouterNotFound
 from .google import Google
 from .graphhopper import Graphhopper
+from .ign import IGN
 from .mapbox_osrm import MapboxOSRM
 from .openrouteservice import ORS
 from .opentripplanner_v2 import OpenTripPlannerV2
@@ -28,6 +29,7 @@ from .valhalla import Valhalla
 _SERVICE_TO_ROUTER = {
     "google": Google,
     "graphhopper": Graphhopper,
+    "ign": IGN,
     "mapbox_osrm": MapboxOSRM,
     "mapbox-osrm": MapboxOSRM,
     "mapbox": MapboxOSRM,
@@ -59,7 +61,7 @@ def get_router_by_name(router_name):
     :param router_name: Name of the router as string.
     :type router_name: str
 
-    :rtype: Union[:class:`routingpy.routers.google.Google`, :class:`routingpy.routers.graphhopper.Graphhopper`, :class:`routingpy.routers.mapbox_osrm.MapBoxOSRM`, :class:`routingpy.routers.openrouteservice.ORS`, :class:`routingpy.routers.osrm.OSRM`, :class:`routingpy.routers.otp_v2.OpenTripPlannerV2`, :class:`routingpy.routers.valhalla.Valhalla`]
+    :rtype: Union[:class:`routingpy.routers.google.Google`, :class:`routingpy.routers.graphhopper.Graphhopper`, :class:`routingpy.routers.ign.IGN`, :class:`routingpy.routers.mapbox_osrm.MapBoxOSRM`, :class:`routingpy.routers.openrouteservice.ORS`, :class:`routingpy.routers.osrm.OSRM`, :class:`routingpy.routers.otp_v2.OpenTripPlannerV2`, :class:`routingpy.routers.valhalla.Valhalla`]
 
     """
     try: