+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added `geotiff` support for Valhalla's isochrones ([#150](https://github.com/mthh/routingpy/pull/150)).
 - Added `alternates` support for Valhalla's directions ([#152](https://github.com/mthh/routingpy/pull/152)).
 - Added `/optimized_route` endpoint to Valhalla ([#160](https://github.com/mthh/routingpy/pull/160)).
+- Added `IGN` router with support for `directions` and `isochrones` for French territories ([#157](https://github.com/mthh/routingpy/pull/157)).
 
 ### Fixed